New wall correction + PID values

+++ Rat/Core/Inc/controller.h
@@ -7,7 +7,7 @@
 
 #include "main.h"
 
-void move(int8_t n);
+void move(float n);
 void turn(int n);
 
 #endif /* INC_CONTROLLER_H_ */
