+++ Rat/Core/Src/controller.c
@@ -12,7 +12,7 @@ int mazeSquare = 100;
 /*
  * We recommend you implement this function so that move(1) will move your rat 1 cell forward.
  */
-void move(int8_t n) {
+void move(float n) {
 	/*
 	 * For assignment 3.1: Don't worry about implementing this yet
 	 * For assignment 3.2: This function should set the distance and angle goals appropriately for PID (hint: using the setGoal functions in pid.c)
@@ -22,7 +22,7 @@ void move(int8_t n) {
 	 *
 	 * You should also call resetPID before exiting this function so your rat is ready for the next instruction.
 	 */
-	setPIDGoalD(n * mazeSquare);
+	setPIDGoalD((int)(n * mazeSquare));
 	while (PIDdone() == 0){
 		updatePID();
 		loopct++;
@@ -52,7 +52,7 @@ void turn(int n) {
 	}
 	resetPID();
 }
-/*
+
 void moveAndTurn(int8_t nM, int nT){
 	setPIDGoalD(nM * mazeSquare);
 	setPIDGoalA(nT);
@@ -62,4 +62,4 @@ void moveAndTurn(int8_t nM, int nT){
 		}
 		resetPID();
 }
-*/
+
